everything works except icons
