+++ code/__HELPERS/_lists.dm
@@ -731,3 +731,7 @@
 			return FALSE
 
 	return TRUE
+
+///sort any value in a list
+/proc/sort_list(list/list_to_sort, cmp=/proc/cmp_text_asc)
+	return sortTim(list_to_sort.Copy(), cmp)
