+++ code/__HELPERS/global_lists.dm
@@ -69,7 +69,59 @@
 			continue
 		GLOB.skillcapes[A.id] = A
 
-	init_subtypes(/datum/crafting_recipe, GLOB.crafting_recipes)
+	init_crafting_recipes()
+	init_crafting_recipes_atoms()
+
+/// Inits crafting recipe lists
+/proc/init_crafting_recipes(list/crafting_recipes)
+	for(var/path in subtypesof(/datum/crafting_recipe))
+		var/is_cooking = ispath(path, /datum/crafting_recipe/food/)
+		var/datum/crafting_recipe/recipe = new path()
+		recipe.reqs = sort_list(recipe.reqs, GLOBAL_PROC_REF(cmp_crafting_req_priority))
+		if(recipe.name != "" && recipe.result)
+			if(is_cooking)
+				GLOB.cooking_recipes += recipe
+			else
+				GLOB.crafting_recipes += recipe
+
+/// Inits atoms used in crafting recipes
+/proc/init_crafting_recipes_atoms()
+	var/list/recipe_lists = list(
+		GLOB.crafting_recipes,
+		GLOB.cooking_recipes
+	)
+	var/list/atom_lists = list(
+		GLOB.crafting_recipes_atoms,
+		GLOB.cooking_recipes_atoms
+	)
+
+	for(var/recipe_list in recipe_lists)
+		for(var/datum/crafting_recipe/recipe as anything in recipe_list)
+			var/list_index = recipe_lists.Find(recipe_list)
+			// Result
+			if(!(recipe.result in atom_lists[list_index]))
+				atom_lists[list_index] += recipe.result
+			// Ingredients
+			for(var/atom/req_atom as anything in recipe.reqs)
+				if(!(req_atom in atom_lists[list_index]))
+					atom_lists[list_index] += req_atom
+			// Catalysts
+			for(var/atom/req_atom as anything in recipe.chem_catalysts)
+				if(!(req_atom in atom_lists[list_index]))
+					atom_lists[list_index] += req_atom
+			// Reaction data - required container
+			if(recipe.reaction)
+				var/required_container = initial(recipe.reaction.required_container)
+				if(required_container && !(required_container in atom_lists[list_index]))
+					atom_lists[list_index] += required_container
+			// Tools
+			for(var/atom/req_atom as anything in recipe.tool_paths)
+				if(!(req_atom in atom_lists[list_index]))
+					atom_lists[list_index] += req_atom
+			// Machinery
+			for(var/atom/req_atom as anything in recipe.machinery)
+				if(!(req_atom in atom_lists[list_index]))
+					atom_lists[list_index] += req_atom
 
 //creates every subtype of prototype (excluding prototype) and adds it to list L.
 //if no list/L is provided, one is created.
