+++ code/_globalvars/lists/objects.dm
@@ -18,7 +18,10 @@ GLOBAL_LIST(chemical_reactions_list)				//list of all /datum/chemical_reaction d
 GLOBAL_LIST(chemical_reagents_list)				//list of all /datum/reagent datums indexed by reagent id. Used by chemistry stuff
 GLOBAL_LIST_EMPTY(tech_list)					//list of all /datum/tech datums indexed by id.
 GLOBAL_LIST_EMPTY(surgeries_list)				//list of all surgeries by name, associated with their path.
-GLOBAL_LIST_EMPTY(crafting_recipes)				//list of all table craft recipes
+GLOBAL_LIST_EMPTY(crafting_recipes) //list of all crafting recipes
+GLOBAL_LIST_EMPTY(crafting_recipes_atoms) //list of all atoms occuring in crafting recipes
+GLOBAL_LIST_EMPTY(cooking_recipes) //list of all cooking recipes
+GLOBAL_LIST_EMPTY(cooking_recipes_atoms) //list of all atoms occuring in cooking recipes
 GLOBAL_LIST_EMPTY(rcd_list)					//list of Rapid Construction Devices.
 GLOBAL_LIST_EMPTY(apcs_list)					//list of all Area Power Controller machines, separate from machines for powernet speeeeeeed.
 GLOBAL_LIST_EMPTY(tracked_implants)			//list of all current implants that are tracked to work out what sort of trek everyone is on. Sadly not on lavaworld not implemented...
