+++ code/modules/antagonists/bloodsuckers/structures/bloodsucker_recipes.dm
@@ -169,7 +169,7 @@
 	reqs = list(/obj/item/stack/sheet/mineral/wood = 3)
 	time = 8 SECONDS
 	category = CAT_WEAPONRY
-	subcategory = CAT_WEAPON
+	category = CAT_WEAPON
 
 /datum/crafting_recipe/woodenducky
 	name = "Wooden Ducky"
@@ -181,7 +181,7 @@
 	)
 	time = 6 SECONDS
 	category = CAT_WEAPONRY
-	subcategory = CAT_WEAPON
+	category = CAT_WEAPON
 	always_available = FALSE
 
 /datum/crafting_recipe/hardened_stake
@@ -191,7 +191,7 @@
 	reqs = list(/obj/item/stack/rods = 1)
 	time = 6 SECONDS
 	category = CAT_WEAPONRY
-	subcategory = CAT_WEAPON
+	category = CAT_WEAPON
 	always_available = FALSE
 
 /datum/crafting_recipe/silver_stake
@@ -204,7 +204,7 @@
 	)
 	time = 8 SECONDS
 	category = CAT_WEAPONRY
-	subcategory = CAT_WEAPON
+	category = CAT_WEAPON
 	always_available = FALSE
 
 ////////////////////////
