+++ code/modules/food_and_drinks/recipes/tablecraft/recipes_bread.dm
@@ -11,7 +11,7 @@
 				/obj/item/reagent_containers/food/snacks/pastrybase = 2
 	)
 	result = /obj/item/reagent_containers/food/snacks/baguette
-	subcategory = CAT_BREAD
+	category = CAT_BREAD
 
 /datum/crafting_recipe/food/banananutbread
 	name = "Banana-Nut Bread"
@@ -22,7 +22,7 @@
 		/obj/item/reagent_containers/food/snacks/grown/banana = 1
 	)
 	result = /obj/item/reagent_containers/food/snacks/store/bread/banana
-	subcategory = CAT_BREAD
+	category = CAT_BREAD
 
 /datum/crafting_recipe/food/butterbiscuit
 	name = "Butter Biscuit"
@@ -31,7 +31,7 @@
 		/obj/item/reagent_containers/food/snacks/butterslice = 1
 	)
 	result = /obj/item/reagent_containers/food/snacks/butterbiscuit
-	subcategory = CAT_BREAD
+	category = CAT_BREAD
 
 /datum/crafting_recipe/food/butteredtoast
 	name = "Buttered Toast"
@@ -40,7 +40,7 @@
 		/obj/item/reagent_containers/food/snacks/butterslice = 1
 	)
 	result = /obj/item/reagent_containers/food/snacks/butteredtoast
-	subcategory = CAT_BREAD
+	category = CAT_BREAD
 
 /datum/crafting_recipe/food/creamcheesebread
 	name = "Cream Cheese Bread"
@@ -50,7 +50,7 @@
 		/obj/item/reagent_containers/food/snacks/cheesewedge = 2
 	)
 	result = /obj/item/reagent_containers/food/snacks/store/bread/creamcheese
-	subcategory = CAT_BREAD
+	category = CAT_BREAD
 
 /datum/crafting_recipe/food/frenchtoast
 	name = "raw French toast"
@@ -62,7 +62,7 @@
 		/datum/reagent/consumable/eggyolk = 5
 	)
 	result = /obj/item/reagent_containers/food/snacks/frenchtoast/raw
-	subcategory = CAT_BREAD
+	category = CAT_BREAD
 
 /datum/crafting_recipe/food/garlicbread
 	name = "Garlic Bread"
@@ -72,7 +72,7 @@
 				/obj/item/reagent_containers/food/snacks/butterslice = 1
 	)
 	result = /obj/item/reagent_containers/food/snacks/garlicbread
-	subcategory = CAT_BREAD
+	category = CAT_BREAD
 
 /datum/crafting_recipe/food/jelliedyoast
 	name = "Jellied Toast"
@@ -81,7 +81,7 @@
 		/obj/item/reagent_containers/food/snacks/breadslice/toast = 1
 	)
 	result = /obj/item/reagent_containers/food/snacks/jelliedtoast/cherry
-	subcategory = CAT_BREAD
+	category = CAT_BREAD
 
 /datum/crafting_recipe/food/meatbread
 	name = "Meat Bread"
@@ -91,7 +91,7 @@
 		/obj/item/reagent_containers/food/snacks/cheesewedge = 3
 	)
 	result = /obj/item/reagent_containers/food/snacks/store/bread/meat
-	subcategory = CAT_BREAD
+	category = CAT_BREAD
 
 /datum/crafting_recipe/food/mimanabread
 	name = "Mimana Bread"
@@ -102,7 +102,7 @@
 		/obj/item/reagent_containers/food/snacks/grown/banana/mime = 1
 	)
 	result = /obj/item/reagent_containers/food/snacks/store/bread/mimana
-	subcategory = CAT_BREAD
+	category = CAT_BREAD
 
 /datum/crafting_recipe/food/slimetoast
 	name = "Slime Toast"
@@ -111,7 +111,7 @@
 		/obj/item/reagent_containers/food/snacks/breadslice/toast = 1
 	)
 	result = /obj/item/reagent_containers/food/snacks/jelliedtoast/slime
-	subcategory = CAT_BREAD
+	category = CAT_BREAD
 
 /datum/crafting_recipe/food/spidermeatbread
 	name = "Spidermeat Bread"
@@ -121,7 +121,7 @@
 		/obj/item/reagent_containers/food/snacks/cheesewedge = 3
 	)
 	result = /obj/item/reagent_containers/food/snacks/store/bread/spidermeat
-	subcategory = CAT_BREAD
+	category = CAT_BREAD
 
 /datum/crafting_recipe/food/tofubread
 	name = "Tofu Bread"
@@ -131,7 +131,7 @@
 		/obj/item/reagent_containers/food/snacks/cheesewedge = 3
 	)
 	result = /obj/item/reagent_containers/food/snacks/store/bread/tofu
-	subcategory = CAT_BREAD
+	category = CAT_BREAD
 
 /datum/crafting_recipe/food/twobread
 	name = "Two Bread"
@@ -140,7 +140,7 @@
 		/obj/item/reagent_containers/food/snacks/breadslice/plain = 2
 	)
 	result = /obj/item/reagent_containers/food/snacks/twobread
-	subcategory = CAT_BREAD
+	category = CAT_BREAD
 
 /datum/crafting_recipe/food/xenomeatbread
 	name = "Xenomeat Bread"
@@ -150,6 +150,6 @@
 		/obj/item/reagent_containers/food/snacks/cheesewedge = 3
 	)
 	result = /obj/item/reagent_containers/food/snacks/store/bread/xenomeat
-	subcategory = CAT_BREAD
+	category = CAT_BREAD
