+++ code/modules/food_and_drinks/recipes/tablecraft/recipes_cake.dm
@@ -10,7 +10,7 @@
 		/obj/item/reagent_containers/food/snacks/grown/apple = 2
 	)
 	result = /obj/item/reagent_containers/food/snacks/store/cake/apple
-	subcategory = CAT_CAKE
+	category = CAT_CAKE
 
 /datum/crafting_recipe/food/birthdaycake
 	name = "Birthday Cake"
@@ -19,7 +19,7 @@
 		/obj/item/reagent_containers/food/snacks/store/cake/plain = 1
 	)
 	result = /obj/item/reagent_containers/food/snacks/store/cake/birthday
-	subcategory = CAT_CAKE
+	category = CAT_CAKE
 
 /datum/crafting_recipe/food/braincake
 	name = "Brain Cake"
@@ -28,7 +28,7 @@
 		/obj/item/reagent_containers/food/snacks/store/cake/plain = 1
 	)
 	result = /obj/item/reagent_containers/food/snacks/store/cake/brain
-	subcategory = CAT_CAKE
+	category = CAT_CAKE
 
 /datum/crafting_recipe/food/carrotcake
 	name = "Carrot Cake"
@@ -37,7 +37,7 @@
 		/obj/item/reagent_containers/food/snacks/grown/carrot = 2
 	)
 	result = /obj/item/reagent_containers/food/snacks/store/cake/carrot
-	subcategory = CAT_CAKE
+	category = CAT_CAKE
 
 /datum/crafting_recipe/food/cheesecake
 	name = "Cheese Cake"
@@ -46,7 +46,7 @@
 		/obj/item/reagent_containers/food/snacks/cheesewedge = 2
 	)
 	result = /obj/item/reagent_containers/food/snacks/store/cake/cheese
-	subcategory = CAT_CAKE
+	category = CAT_CAKE
 
 /datum/crafting_recipe/food/chocolatecake
 	name = "Chocolate Cake"
@@ -55,7 +55,7 @@
 		/obj/item/reagent_containers/food/snacks/chocolatebar = 2
 	)
 	result = /obj/item/reagent_containers/food/snacks/store/cake/chocolate
-	subcategory = CAT_CAKE
+	category = CAT_CAKE
 
 /datum/crafting_recipe/food/chocolatecheesecake
 	name = "Chocolate Cheese Cake"
@@ -64,7 +64,7 @@
 		/obj/item/reagent_containers/food/snacks/cheesewedge = 2
 	)
 	result = /obj/item/reagent_containers/food/snacks/store/cake/chocolatecheese
-	subcategory = CAT_CAKE
+	category = CAT_CAKE
 
 /datum/crafting_recipe/food/donkcake
 	name = "Donk Cake"
@@ -73,7 +73,7 @@
 		/obj/item/reagent_containers/food/snacks/store/cake/plain = 1
 	)
 	result = /obj/item/reagent_containers/food/snacks/store/cake/donk
-	subcategory = CAT_CAKE
+	category = CAT_CAKE
 
 /datum/crafting_recipe/food/lemoncake
 	name = "Lemon Cake"
@@ -82,7 +82,7 @@
 		/obj/item/reagent_containers/food/snacks/grown/citrus/lemon = 2
 	)
 	result = /obj/item/reagent_containers/food/snacks/store/cake/lemon
-	subcategory = CAT_CAKE
+	category = CAT_CAKE
 
 /datum/crafting_recipe/food/limecake
 	name = "Lime Cake"
@@ -91,7 +91,7 @@
 		/obj/item/reagent_containers/food/snacks/grown/citrus/lime = 2
 	)
 	result = /obj/item/reagent_containers/food/snacks/store/cake/lime
-	subcategory = CAT_CAKE
+	category = CAT_CAKE
 
 /datum/crafting_recipe/food/cak
 	name = "Living Cat/Cake Hybrid"
@@ -105,7 +105,7 @@
 		/datum/reagent/teslium = 1 //To shock the whole thing into life
 	)
 	result = /mob/living/simple_animal/pet/cat/cak
-	subcategory = CAT_CAKE //Cat! Haha, get it? CAT? GET IT???
+	category = CAT_CAKE //Cat! Haha, get it? CAT? GET IT???
 
 /datum/crafting_recipe/food/orangecake
 	name = "Orange Cake"
@@ -114,7 +114,7 @@
 		/obj/item/reagent_containers/food/snacks/grown/citrus/orange = 2
 	)
 	result = /obj/item/reagent_containers/food/snacks/store/cake/orange
-	subcategory = CAT_CAKE
+	category = CAT_CAKE
 
 /datum/crafting_recipe/food/pumpkinspicecake
 	name = "Pumpkin Spice Cake"
@@ -123,7 +123,7 @@
 		/obj/item/reagent_containers/food/snacks/grown/pumpkin = 2
 	)
 	result = /obj/item/reagent_containers/food/snacks/store/cake/pumpkinspice
-	subcategory = CAT_CAKE
+	category = CAT_CAKE
 
 /datum/crafting_recipe/food/slimecake
 	name = "Slime Cake"
@@ -132,7 +132,7 @@
 		/obj/item/reagent_containers/food/snacks/store/cake/plain = 1
 	)
 	result = /obj/item/reagent_containers/food/snacks/store/cake/slimecake
-	subcategory = CAT_CAKE
+	category = CAT_CAKE
 
 /datum/crafting_recipe/food/weddingcake
 	name = "Wedding Cake"
@@ -141,4 +141,4 @@
 		/datum/reagent/consumable/sugar = 100
 	)
 	result = /obj/item/reagent_containers/food/snacks/store/cake/wedding
-	subcategory = CAT_CAKE
+	category = CAT_CAKE
