+++ code/modules/food_and_drinks/recipes/tablecraft/recipes_egg.dm
@@ -10,7 +10,7 @@
 		/obj/item/reagent_containers/food/snacks/chocolatebar = 1
 	)
 	result = /obj/item/reagent_containers/food/snacks/chocolateegg
-	subcategory = CAT_EGG
+	category = CAT_EGG
 
 /datum/crafting_recipe/food/eggsbenedict
 	name = "Eggs Benedict"
@@ -20,7 +20,7 @@
 		/obj/item/reagent_containers/food/snacks/breadslice/plain = 1
 	)
 	result = /obj/item/reagent_containers/food/snacks/benedict
-	subcategory = CAT_EGG
+	category = CAT_EGG
 
 /datum/crafting_recipe/food/eggbowl
 	name = "Egg Bowl"
@@ -31,7 +31,7 @@
 		/obj/item/reagent_containers/food/snacks/grown/corn = 1
 	)
 	result = /obj/item/reagent_containers/food/snacks/salad/eggbowl
-	subcategory = CAT_EGG
+	category = CAT_EGG
 
 /datum/crafting_recipe/food/wrap
 	name = "Egg Wrap"
@@ -41,7 +41,7 @@
 		/obj/item/reagent_containers/food/snacks/grown/cabbage = 1
 	)
 	result = /obj/item/reagent_containers/food/snacks/eggwrap
-	subcategory = CAT_EGG
+	category = CAT_EGG
 
 /datum/crafting_recipe/food/friedegg
 	name = "Fried Egg"
@@ -51,7 +51,7 @@
 		/obj/item/reagent_containers/food/snacks/egg = 1
 	)
 	result = /obj/item/reagent_containers/food/snacks/friedegg
-	subcategory = CAT_EGG
+	category = CAT_EGG
 
 /datum/crafting_recipe/food/spidereggsham
 	name = "Green Eggs and Ham"
@@ -61,7 +61,7 @@
 		/obj/item/reagent_containers/food/snacks/meat/cutlet/spider = 2
 	)
 	result = /obj/item/reagent_containers/food/snacks/spidereggsham
-	subcategory = CAT_EGG
+	category = CAT_EGG
 
 /datum/crafting_recipe/food/eggdog
 	name = "Living Egg/Dog Hybrid"
@@ -75,7 +75,7 @@
 		/datum/reagent/teslium = 1 //To shock the whole thing into life
 	)
 	result = /mob/living/simple_animal/pet/dog/eggdog
-	subcategory = CAT_EGG
+	category = CAT_EGG
 
 /datum/crafting_recipe/food/omelette
 	name = "Omelette Du Fromage"
@@ -84,4 +84,4 @@
 		/obj/item/reagent_containers/food/snacks/cheesewedge = 2
 	)
 	result = /obj/item/reagent_containers/food/snacks/omelette
-	subcategory = CAT_EGG
+	category = CAT_EGG
