+++ code/modules/food_and_drinks/recipes/tablecraft/recipes_frozen.dm
@@ -11,7 +11,7 @@
 		/obj/item/reagent_containers/food/snacks/icecream = 1
 	)
 	result = /obj/item/reagent_containers/food/snacks/icecreamsandwich
-	subcategory = CAT_ICE
+	category = CAT_ICE
 
 /datum/crafting_recipe/food/spacefreezy
 	name ="Space freezy"
@@ -21,7 +21,7 @@
 		/obj/item/reagent_containers/food/snacks/icecream = 1
 	)
 	result = /obj/item/reagent_containers/food/snacks/spacefreezy
-	subcategory = CAT_ICE
+	category = CAT_ICE
 
 /datum/crafting_recipe/food/sundae
 	name ="Sundae"
@@ -32,7 +32,7 @@
 		/obj/item/reagent_containers/food/snacks/icecream = 1
 	)
 	result = /obj/item/reagent_containers/food/snacks/sundae
-	subcategory = CAT_ICE
+	category = CAT_ICE
 
 /datum/crafting_recipe/food/honkdae
 	name ="Honkdae"
@@ -44,7 +44,7 @@
 		/obj/item/reagent_containers/food/snacks/icecream = 1
 	)
 	result = /obj/item/reagent_containers/food/snacks/honkdae
-	subcategory = CAT_ICE
+	category = CAT_ICE
 
 /datum/crafting_recipe/food/taiyaki
 	name ="Vanilla Taiyaki"
@@ -55,7 +55,7 @@
 		/datum/reagent/consumable/cooking_oil/fish = 1
 	)
 	result = /obj/item/reagent_containers/food/snacks/taiyaki
-	subcategory = CAT_ICE
+	category = CAT_ICE
 
 /datum/crafting_recipe/food/taiyaki/chocolate
 	name ="Chocolate Taiyaki"
@@ -67,7 +67,7 @@
 		/datum/reagent/consumable/coco = 2
 	)
 	result = /obj/item/reagent_containers/food/snacks/taiyaki/chocolate
-	subcategory = CAT_ICE
+	category = CAT_ICE
 
 /datum/crafting_recipe/food/taiyaki/strawberry
 	name ="Strawberry Taiyaki"
@@ -79,7 +79,7 @@
 		/datum/reagent/consumable/berryjuice = 2
 	)
 	result = /obj/item/reagent_containers/food/snacks/taiyaki/strawberry
-	subcategory = CAT_ICE
+	category = CAT_ICE
 
 /datum/crafting_recipe/food/taiyaki/blue
 	name ="Blue Taiyaki"
@@ -91,7 +91,7 @@
 		/datum/reagent/consumable/ethanol/singulo = 2
 	)
 	result = /obj/item/reagent_containers/food/snacks/taiyaki/blue
-	subcategory = CAT_ICE
+	category = CAT_ICE
 
 /datum/crafting_recipe/food/taiyaki/mobflavor
 	name ="Blood Taiyaki"
@@ -103,7 +103,7 @@
 		/datum/reagent/blood = 2
 	)
 	result = /obj/item/reagent_containers/food/snacks/taiyaki/mobflavor
-	subcategory = CAT_ICE
+	category = CAT_ICE
 
 //////////////////////////SNOW CONES///////////////////////
 
@@ -115,7 +115,7 @@
 		/datum/reagent/consumable/ice = 15
 	)
 	result = /obj/item/reagent_containers/food/snacks/snowcones
-	subcategory = CAT_ICE
+	category = CAT_ICE
 
 /datum/crafting_recipe/food/pineapple_sc
 	name = "Pineapple snowcone"
@@ -126,7 +126,7 @@
 		/obj/item/reagent_containers/food/snacks/pineappleslice = 2
 	)
 	result = /obj/item/reagent_containers/food/snacks/snowcones/pineapple
-	subcategory = CAT_ICE
+	category = CAT_ICE
 
 /datum/crafting_recipe/food/lime_sc
 	name = "Lime snowcone"
@@ -137,7 +137,7 @@
 		/datum/reagent/consumable/limejuice = 5
 	)
 	result = /obj/item/reagent_containers/food/snacks/snowcones/lime
-	subcategory = CAT_ICE
+	category = CAT_ICE
 
 /datum/crafting_recipe/food/lemon_sc
 	name = "Lemon snowcone"
@@ -148,7 +148,7 @@
 		/datum/reagent/consumable/lemonjuice = 5
 	)
 	result = /obj/item/reagent_containers/food/snacks/snowcones/lemon
-	subcategory = CAT_ICE
+	category = CAT_ICE
 
 /datum/crafting_recipe/food/apple_sc
 	name = "Apple snowcone"
@@ -159,7 +159,7 @@
 		/datum/reagent/consumable/applejuice = 5
 	)
 	result = /obj/item/reagent_containers/food/snacks/snowcones/apple
-	subcategory = CAT_ICE
+	category = CAT_ICE
 
 /datum/crafting_recipe/food/grape_sc
 	name = "Grape snowcone"
@@ -170,7 +170,7 @@
 		/datum/reagent/consumable/berryjuice = 5
 	)
 	result = /obj/item/reagent_containers/food/snacks/snowcones/grape
-	subcategory = CAT_ICE
+	category = CAT_ICE
 
 /datum/crafting_recipe/food/orange_sc
 	name = "Orange snowcone"
@@ -181,7 +181,7 @@
 		/datum/reagent/consumable/orangejuice = 5
 	)
 	result = /obj/item/reagent_containers/food/snacks/snowcones/orange
-	subcategory = CAT_ICE
+	category = CAT_ICE
 
 /datum/crafting_recipe/food/blue_sc
 	name = "Bluecherry snowcone"
@@ -192,7 +192,7 @@
 		/datum/reagent/consumable/bluecherryjelly= 5
 	)
 	result = /obj/item/reagent_containers/food/snacks/snowcones/blue
-	subcategory = CAT_ICE
+	category = CAT_ICE
 
 /datum/crafting_recipe/food/red_sc
 	name = "Cherry snowcone"
@@ -203,7 +203,7 @@
 		/datum/reagent/consumable/cherryjelly= 5
 	)
 	result = /obj/item/reagent_containers/food/snacks/snowcones/red
-	subcategory = CAT_ICE
+	category = CAT_ICE
 
 /datum/crafting_recipe/food/mix_sc
 	name = "Mixed berry snowcone"
@@ -214,7 +214,7 @@
 		/datum/reagent/consumable/berryjuice = 15
 	)
 	result = /obj/item/reagent_containers/food/snacks/snowcones/mix
-	subcategory = CAT_ICE
+	category = CAT_ICE
 
 /datum/crafting_recipe/food/fruitsalad_sc
 	name = "Fruit salad snowcone"
@@ -227,7 +227,7 @@
 		/datum/reagent/consumable/lemonjuice = 5
 	)
 	result = /obj/item/reagent_containers/food/snacks/snowcones/fruitsalad
-	subcategory = CAT_ICE
+	category = CAT_ICE
 
 /datum/crafting_recipe/food/mime_sc
 	name = "Mime snowcone"
@@ -238,7 +238,7 @@
 		/datum/reagent/consumable/nothing = 5
 	)
 	result = /obj/item/reagent_containers/food/snacks/snowcones/mime
-	subcategory = CAT_ICE
+	category = CAT_ICE
 
 /datum/crafting_recipe/food/clown_sc
 	name = "Clown snowcone"
@@ -249,7 +249,7 @@
 		/datum/reagent/consumable/clownstears = 5
 	)
 	result = /obj/item/reagent_containers/food/snacks/snowcones/clown
-	subcategory = CAT_ICE
+	category = CAT_ICE
 
 /datum/crafting_recipe/food/soda_sc
 	name = "Soda water snowcone"
@@ -260,7 +260,7 @@
 		/datum/reagent/consumable/sodawater = 15
 	)
 	result = /obj/item/reagent_containers/food/snacks/snowcones/soda
-	subcategory = CAT_ICE
+	category = CAT_ICE
 
 /datum/crafting_recipe/food/pwgrmer_sc
 	name = "Pwergamer snowcone"
@@ -271,7 +271,7 @@
 		/datum/reagent/consumable/pwr_game = 15
 	)
 	result = /obj/item/reagent_containers/food/snacks/snowcones/pwgrmer
-	subcategory = CAT_ICE
+	category = CAT_ICE
 
 /datum/crafting_recipe/food/honey_sc
 	name = "Honey snowcone"
@@ -282,7 +282,7 @@
 		/datum/reagent/consumable/honey = 5
 	)
 	result = /obj/item/reagent_containers/food/snacks/snowcones/honey
-	subcategory = CAT_ICE
+	category = CAT_ICE
 
 /datum/crafting_recipe/food/rainbow_sc
 	name = "Rainbow snowcone"
@@ -293,4 +293,4 @@
 		/datum/reagent/colorful_reagent = 1 //Harder to make
 	)
 	result = /obj/item/reagent_containers/food/snacks/snowcones/rainbow
-	subcategory = CAT_ICE
+	category = CAT_ICE
