+++ code/modules/food_and_drinks/recipes/tablecraft/recipes_meat.dm
@@ -9,7 +9,7 @@
 		/obj/item/reagent_containers/food/snacks/deadmouse = 2
 	)
 	result = /obj/item/reagent_containers/food/snacks/kebab/rat/double
-	subcategory = CAT_MEAT
+	category = CAT_MEAT
 
 /datum/crafting_recipe/food/humankebab
 	name = "Human Kebab"
@@ -18,7 +18,7 @@
 		/obj/item/reagent_containers/food/snacks/meat/steak/plain/human = 2
 	)
 	result = /obj/item/reagent_containers/food/snacks/kebab/human
-	subcategory = CAT_MEAT
+	category = CAT_MEAT
 
 /datum/crafting_recipe/food/kebab
 	name = "Kebab"
@@ -27,7 +27,7 @@
 		/obj/item/reagent_containers/food/snacks/meat/steak = 2
 	)
 	result = /obj/item/reagent_containers/food/snacks/kebab/monkey
-	subcategory = CAT_MEAT
+	category = CAT_MEAT
 
 /datum/crafting_recipe/food/tailkebab
 	name = "Lizard Tail Kebab"
@@ -36,7 +36,7 @@
 		/obj/item/organ/tail/lizard = 1
 	)
 	result = /obj/item/reagent_containers/food/snacks/kebab/tail
-	subcategory = CAT_MEAT
+	category = CAT_MEAT
 
 /datum/crafting_recipe/food/ratkebab
 	name = "Rat Kebab"
@@ -45,7 +45,7 @@
 		/obj/item/reagent_containers/food/snacks/deadmouse = 1
 	)
 	result = /obj/item/reagent_containers/food/snacks/kebab/rat
-	subcategory = CAT_MEAT
+	category = CAT_MEAT
 
 /datum/crafting_recipe/food/tofukebab
 	name = "Tofu Kebab"
@@ -54,7 +54,7 @@
 		/obj/item/reagent_containers/food/snacks/tofu = 2
 	)
 	result = /obj/item/reagent_containers/food/snacks/kebab/tofu
-	subcategory = CAT_MEAT
+	category = CAT_MEAT
 
 ////////////////////////////////////////////////OTHER////////////////////////////////////////////////
 
@@ -65,7 +65,7 @@
 		/obj/item/reagent_containers/food/snacks/meat/cutlet = 1
 	)
 	result = /obj/item/reagent_containers/food/snacks/nugget
-	subcategory = CAT_MEAT
+	category = CAT_MEAT
 
 /datum/crafting_recipe/food/cornedbeef
 	name = "Corned Beef and Cabbage"
@@ -75,7 +75,7 @@
 		/obj/item/reagent_containers/food/snacks/grown/cabbage = 2
 	)
 	result = /obj/item/reagent_containers/food/snacks/cornedbeef
-	subcategory = CAT_MEAT
+	category = CAT_MEAT
 
 /datum/crafting_recipe/food/enchiladas
 	name = "Enchiladas"
@@ -85,7 +85,7 @@
 		/obj/item/reagent_containers/food/snacks/tortilla = 2
 	)
 	result = /obj/item/reagent_containers/food/snacks/enchiladas
-	subcategory = CAT_MEAT
+	category = CAT_MEAT
 
 /datum/crafting_recipe/food/bearsteak
 	name = "Filet Migrawr"
@@ -95,7 +95,7 @@
 	)
 	tool_paths = list(/obj/item/lighter)
 	result = /obj/item/reagent_containers/food/snacks/bearsteak
-	subcategory = CAT_MEAT
+	category = CAT_MEAT
 
 /datum/crafting_recipe/food/pigblanket
 	name = "Pig in a Blanket"
@@ -105,7 +105,7 @@
 		/obj/item/reagent_containers/food/snacks/meat/cutlet = 1
 	)
 	result = /obj/item/reagent_containers/food/snacks/pigblanket
-	subcategory = CAT_MEAT
+	category = CAT_MEAT
 
 /datum/crafting_recipe/food/rawkhinkali
 	name = "Raw Khinkali"
@@ -114,7 +114,7 @@
 		/obj/item/reagent_containers/food/snacks/meatball = 1
 	)
 	result =  /obj/item/reagent_containers/food/snacks/rawkhinkali
-	subcategory = CAT_MEAT
+	category = CAT_MEAT
 
 /datum/crafting_recipe/food/ricepork
 	name = "Rice and Pork"
@@ -123,7 +123,7 @@
 		/obj/item/reagent_containers/food/snacks/meat/cutlet = 2
 	)
 	result = /obj/item/reagent_containers/food/snacks/salad/ricepork
-	subcategory = CAT_MEAT
+	category = CAT_MEAT
 
 /datum/crafting_recipe/food/sausage
 	name = "Raw Sausage"
@@ -132,7 +132,7 @@
 		/obj/item/reagent_containers/food/snacks/meat/raw_cutlet = 2
 	)
 	result = /obj/item/reagent_containers/food/snacks/raw_sausage
-	subcategory = CAT_MEAT
+	category = CAT_MEAT
 
 /datum/crafting_recipe/food/stewedsoymeat
 	name = "Stewed Soymeat"
@@ -142,7 +142,7 @@
 		/obj/item/reagent_containers/food/snacks/grown/tomato = 1
 	)
 	result = /obj/item/reagent_containers/food/snacks/stewedsoymeat
-	subcategory = CAT_MEAT
+	category = CAT_MEAT
 
 /datum/crafting_recipe/food/meatclown
 	name = "Meat Clown"
@@ -151,7 +151,7 @@
 		/obj/item/reagent_containers/food/snacks/grown/banana = 1
 	)
 	result = /obj/item/reagent_containers/food/snacks/meatclown
-	subcategory = CAT_MEAT
+	category = CAT_MEAT
 
 /datum/crafting_recipe/food/gumbo
 	name = "Black eyed gumbo"
@@ -162,7 +162,7 @@
 		/obj/item/reagent_containers/food/snacks/meat/cutlet = 1
 	)
 	result = /obj/item/reagent_containers/food/snacks/salad/gumbo
-	subcategory = CAT_MEAT
+	category = CAT_MEAT
 
 /datum/crafting_recipe/food/fishfry
 	name = "Fish fry"
@@ -172,4 +172,4 @@
 		/obj/item/reagent_containers/food/snacks/carpmeat/fish/cooked = 1
 	)
 	result = /obj/item/reagent_containers/food/snacks/fishfry
-	subcategory = CAT_MEAT
+	category = CAT_MEAT
