+++ code/modules/food_and_drinks/recipes/tablecraft/recipes_misc.dm
@@ -10,7 +10,7 @@
 		/obj/item/reagent_containers/food/snacks/grown/soybeans = 2
 	)
 	result = /obj/item/reagent_containers/food/snacks/beans
-	subcategory = CAT_MISCFOOD
+	category = CAT_MISCFOOD
 
 /datum/crafting_recipe/food/branrequests
 	name = "Bran Requests Cereal"
@@ -19,7 +19,7 @@
 		/obj/item/reagent_containers/food/snacks/no_raisin = 1
 	)
 	result = /obj/item/reagent_containers/food/snacks/branrequests
-	subcategory = CAT_MISCFOOD
+	category = CAT_MISCFOOD
 
 /datum/crafting_recipe/food/burrito
 	name ="Burrito"
@@ -28,7 +28,7 @@
 		/obj/item/reagent_containers/food/snacks/grown/soybeans = 2
 	)
 	result = /obj/item/reagent_containers/food/snacks/burrito
-	subcategory = CAT_MISCFOOD
+	category = CAT_MISCFOOD
 
 /datum/crafting_recipe/food/fivelayerburrito
 	name ="Five Layer Burrito"
@@ -36,7 +36,7 @@
 		/obj/item/reagent_containers/food/snacks/burrito = 5
 	)
 	result = /obj/item/reagent_containers/food/snacks/fivelayerburrito
-	subcategory = CAT_MISCFOOD
+	category = CAT_MISCFOOD
 
 /datum/crafting_recipe/food/candiedapple
 	name = "Candied Apple"
@@ -45,7 +45,7 @@
 		/obj/item/reagent_containers/food/snacks/grown/apple = 1
 	)
 	result = /obj/item/reagent_containers/food/snacks/candiedapple
-	subcategory = CAT_MISCFOOD
+	category = CAT_MISCFOOD
 
 /datum/crafting_recipe/food/carneburrito
 	name ="Carne de Asada Burrito"
@@ -55,7 +55,7 @@
 		/obj/item/reagent_containers/food/snacks/grown/soybeans = 1
 	)
 	result = /obj/item/reagent_containers/food/snacks/carneburrito
-	subcategory = CAT_MISCFOOD
+	category = CAT_MISCFOOD
 
 /datum/crafting_recipe/food/cheesyburrito
 	name ="Cheesy Burrito"
@@ -65,7 +65,7 @@
 		/obj/item/reagent_containers/food/snacks/grown/soybeans = 1
 	)
 	result = /obj/item/reagent_containers/food/snacks/cheesyburrito
-	subcategory = CAT_MISCFOOD
+	category = CAT_MISCFOOD
 
 /datum/crafting_recipe/food/cheesyfries
 	name = "Cheesy Fries"
@@ -74,7 +74,7 @@
 		/obj/item/reagent_containers/food/snacks/cheesewedge = 1
 	)
 	result = /obj/item/reagent_containers/food/snacks/cheesyfries
-	subcategory = CAT_MISCFOOD
+	category = CAT_MISCFOOD
 
 /datum/crafting_recipe/food/poutine
 	name = "Poutine"
@@ -84,7 +84,7 @@
 		/datum/reagent/consumable/gravy = 10
 	)
 	result = /obj/item/reagent_containers/food/snacks/poutine
-	subcategory = CAT_MISCFOOD
+	category = CAT_MISCFOOD
 
 /datum/crafting_recipe/food/cheesynachos
 	name ="Cheesy Nachos"
@@ -94,7 +94,7 @@
 		/obj/item/reagent_containers/food/snacks/tortilla = 1
 	)
 	result = /obj/item/reagent_containers/food/snacks/cheesynachos
-	subcategory = CAT_MISCFOOD
+	category = CAT_MISCFOOD
 
 /datum/crafting_recipe/food/chococoin
 	name = "Chocolate Coin"
@@ -103,7 +103,7 @@
 		/obj/item/reagent_containers/food/snacks/chocolatebar = 1
 	)
 	result = /obj/item/reagent_containers/food/snacks/chococoin
-	subcategory = CAT_MISCFOOD
+	category = CAT_MISCFOOD
 
 /datum/crafting_recipe/food/chocoorange
 	name = "Chocolate Orange"
@@ -112,7 +112,7 @@
 		/obj/item/reagent_containers/food/snacks/chocolatebar = 1
 	)
 	result = /obj/item/reagent_containers/food/snacks/chocoorange
-	subcategory = CAT_MISCFOOD
+	category = CAT_MISCFOOD
 
 /datum/crafting_recipe/food/taco
 	name ="Classic Taco"
@@ -123,7 +123,7 @@
 		/obj/item/reagent_containers/food/snacks/grown/cabbage = 1
 	)
 	result = /obj/item/reagent_containers/food/snacks/taco
-	subcategory = CAT_MISCFOOD
+	category = CAT_MISCFOOD
 
 /datum/crafting_recipe/food/cubannachos
 	name ="Cuban Nachos"
@@ -133,7 +133,7 @@
 		/obj/item/reagent_containers/food/snacks/tortilla = 1
 	)
 	result = /obj/item/reagent_containers/food/snacks/cubannachos
-	subcategory = CAT_MISCFOOD
+	category = CAT_MISCFOOD
 
 /datum/crafting_recipe/food/eggplantparm
 	name ="Eggplant Parmigiana"
@@ -142,7 +142,7 @@
 		/obj/item/reagent_containers/food/snacks/grown/eggplant = 1
 	)
 	result = /obj/item/reagent_containers/food/snacks/eggplantparm
-	subcategory = CAT_MISCFOOD
+	category = CAT_MISCFOOD
 
 /datum/crafting_recipe/food/fudgedice
 	name = "Fudge Dice"
@@ -151,7 +151,7 @@
 		/obj/item/reagent_containers/food/snacks/chocolatebar = 1
 	)
 	result = /obj/item/reagent_containers/food/snacks/fudgedice
-	subcategory = CAT_MISCFOOD
+	category = CAT_MISCFOOD
 
 /datum/crafting_recipe/food/fuegoburrito
 	name ="Fuego Plasma Burrito"
@@ -161,7 +161,7 @@
 		/obj/item/reagent_containers/food/snacks/grown/soybeans = 1
 	)
 	result = /obj/item/reagent_containers/food/snacks/fuegoburrito
-	subcategory = CAT_MISCFOOD
+	category = CAT_MISCFOOD
 
 /datum/crafting_recipe/food/honeybar
 	name = "Honey Nut Bar"
@@ -170,7 +170,7 @@
 		/datum/reagent/consumable/honey = 5
 	)
 	result = /obj/item/reagent_containers/food/snacks/honeybar
-	subcategory = CAT_MISCFOOD
+	category = CAT_MISCFOOD
 
 /datum/crafting_recipe/food/loadedbakedpotato
 	name = "Loaded Baked Potato"
@@ -180,7 +180,7 @@
 		/obj/item/reagent_containers/food/snacks/cheesewedge = 1
 	)
 	result = /obj/item/reagent_containers/food/snacks/loadedbakedpotato
-	subcategory = CAT_MISCFOOD
+	category = CAT_MISCFOOD
 
 /datum/crafting_recipe/food/melonfruitbowl
 	name ="Melon Fruit Bowl"
@@ -193,7 +193,7 @@
 		/obj/item/reagent_containers/food/snacks/grown/ambrosia = 1
 	)
 	result = /obj/item/reagent_containers/food/snacks/melonfruitbowl
-	subcategory = CAT_MISCFOOD
+	category = CAT_MISCFOOD
 
 /datum/crafting_recipe/food/melonkeg
 	name ="Melon Keg"
@@ -204,7 +204,7 @@
 	)
 	parts = list(/obj/item/reagent_containers/food/drinks/bottle/vodka = 1)
 	result = /obj/item/reagent_containers/food/snacks/melonkeg
-	subcategory = CAT_MISCFOOD
+	category = CAT_MISCFOOD
 
 /datum/crafting_recipe/food/nachos
 	name ="Nachos"
@@ -213,7 +213,7 @@
 		/obj/item/reagent_containers/food/snacks/tortilla = 1
 	)
 	result = /obj/item/reagent_containers/food/snacks/nachos
-	subcategory = CAT_MISCFOOD
+	category = CAT_MISCFOOD
 
 /datum/crafting_recipe/food/tacoplain
 	name ="Plain Taco"
@@ -223,7 +223,7 @@
 		/obj/item/reagent_containers/food/snacks/meat/cutlet = 1
 	)
 	result = /obj/item/reagent_containers/food/snacks/taco/plain
-	subcategory = CAT_MISCFOOD
+	category = CAT_MISCFOOD
 
 /datum/crafting_recipe/food/powercrepe
 	name = "Powercrepe"
@@ -236,7 +236,7 @@
 		/obj/item/melee/sabre = 1
 	)
 	result = /obj/item/reagent_containers/food/snacks/powercrepe
-	subcategory = CAT_MISCFOOD
+	category = CAT_MISCFOOD
 
 /datum/crafting_recipe/food/ricepudding
 	name = "Rice Pudding"
@@ -246,7 +246,7 @@
 		/obj/item/reagent_containers/food/snacks/salad/boiledrice = 1
 	)
 	result = /obj/item/reagent_containers/food/snacks/salad/ricepudding
-	subcategory = CAT_MISCFOOD
+	category = CAT_MISCFOOD
 
 /datum/crafting_recipe/food/spiderlollipop
 	name = "Spider Lollipop"
@@ -256,7 +256,7 @@
 		/obj/item/reagent_containers/food/snacks/spiderling = 1
 	)
 	result = /obj/item/reagent_containers/food/snacks/spiderlollipop
-	subcategory = CAT_MISCFOOD
+	category = CAT_MISCFOOD
 
 /datum/crafting_recipe/food/stuffedlegion
 	name = "Stuffed Legion"
@@ -268,7 +268,7 @@
 		/datum/reagent/consumable/capsaicin = 2
 	)
 	result = /obj/item/reagent_containers/food/snacks/stuffedlegion
-	subcategory = CAT_MISCFOOD
+	category = CAT_MISCFOOD
 
 /datum/crafting_recipe/food/candy_strawberry
 	name = "Strawberry Flavored Candy"
@@ -280,7 +280,7 @@
 	)
 	tool_paths = list(/obj/item/toy/crayon)
 	result = /obj/item/reagent_containers/food/snacks/candy_strawberry
-	subcategory = CAT_MISCFOOD
+	category = CAT_MISCFOOD
 
 
 /datum/crafting_recipe/food/royalcheese
@@ -292,7 +292,7 @@
 		/datum/reagent/toxin/mutagen = 5
 	)
 	result = /obj/item/reagent_containers/food/snacks/royalcheese
-	subcategory = CAT_MISCFOOD
+	category = CAT_MISCFOOD
 
 /datum/crafting_recipe/food/chipsandsalsa
 	name = "Chips and Salsa"
@@ -303,4 +303,4 @@
 		/obj/item/reagent_containers/food/snacks/grown/tomato = 1
 	)
 	result = /obj/item/reagent_containers/food/snacks/chipsandsalsa
-	subcategory = CAT_MISCFOOD
+	category = CAT_MISCFOOD
