+++ code/modules/food_and_drinks/recipes/tablecraft/recipes_pie.dm
@@ -10,7 +10,7 @@
 		/obj/item/reagent_containers/food/snacks/grown/mushroom/amanita = 1
 	)
 	result = /obj/item/reagent_containers/food/snacks/pie/amanita_pie
-	subcategory = CAT_PIE
+	category = CAT_PIE
 
 /datum/crafting_recipe/food/applepie
 	name = "Apple Pie"
@@ -19,7 +19,7 @@
 		/obj/item/reagent_containers/food/snacks/grown/apple = 1
 	)
 	result = /obj/item/reagent_containers/food/snacks/pie/applepie
-	subcategory = CAT_PIE
+	category = CAT_PIE
 
 /datum/crafting_recipe/food/bananacreampie
 	name = "Banana Cream Pie"
@@ -29,7 +29,7 @@
 		 /obj/item/reagent_containers/food/snacks/grown/banana = 1
 	)
 	result = /obj/item/reagent_containers/food/snacks/pie/cream
-	subcategory = CAT_PIE
+	category = CAT_PIE
 
 /datum/crafting_recipe/food/bearypie
 	name = "Beary Pie"
@@ -39,7 +39,7 @@
 		/obj/item/reagent_containers/food/snacks/meat/steak/bear = 1
 	)
 	result = /obj/item/reagent_containers/food/snacks/pie/bearypie
-	subcategory = CAT_PIE
+	category = CAT_PIE
 
 /datum/crafting_recipe/food/berryclafoutis
 	name = "Berry Clafoutis"
@@ -48,7 +48,7 @@
 		/obj/item/reagent_containers/food/snacks/grown/berries = 1
 	)
 	result = /obj/item/reagent_containers/food/snacks/pie/berryclafoutis
-	subcategory = CAT_PIE
+	category = CAT_PIE
 
 /datum/crafting_recipe/food/blumpkinpie
 	name = "Blumpkin Pie"
@@ -59,7 +59,7 @@
 		/obj/item/reagent_containers/food/snacks/grown/blumpkin = 1
 	)
 	result = /obj/item/reagent_containers/food/snacks/pie/blumpkinpie
-	subcategory = CAT_PIE
+	category = CAT_PIE
 
 /datum/crafting_recipe/food/buttcinnpie
 	name = "Butterscotch Cinnamon Pie"
@@ -70,7 +70,7 @@
 		/datum/reagent/consumable/cinnamon = 5
 	)
 	result = /obj/item/reagent_containers/food/snacks/pie/buttcinnpie
-	subcategory = CAT_PIE
+	category = CAT_PIE
 
 /datum/crafting_recipe/food/cherrypie
 	name = "Cherry Pie"
@@ -79,7 +79,7 @@
 		 /obj/item/reagent_containers/food/snacks/grown/cherries = 1
 	)
 	result = /obj/item/reagent_containers/food/snacks/pie/cherrypie
-	subcategory = CAT_PIE
+	category = CAT_PIE
 
 /datum/crafting_recipe/food/dulcedebatata
 	name = "Dulce de Batata"
@@ -89,7 +89,7 @@
 		/obj/item/reagent_containers/food/snacks/grown/potato/sweet = 2
 	)
 	result = /obj/item/reagent_containers/food/snacks/pie/dulcedebatata
-	subcategory = CAT_PIE
+	category = CAT_PIE
 
 /datum/crafting_recipe/food/frostypie
 	name = "Frosty Pie"
@@ -98,7 +98,7 @@
 		/obj/item/reagent_containers/food/snacks/grown/bluecherries = 1
 	)
 	result = /obj/item/reagent_containers/food/snacks/pie/frostypie
-	subcategory = CAT_PIE
+	category = CAT_PIE
 
 /datum/crafting_recipe/food/goldenappletart
 	name = "Golden Apple Streusel Tart"
@@ -109,7 +109,7 @@
 		/obj/item/reagent_containers/food/snacks/grown/apple/gold = 1
 	)
 	result = /obj/item/reagent_containers/food/snacks/pie/appletart
-	subcategory = CAT_PIE
+	category = CAT_PIE
 
 /datum/crafting_recipe/food/grapetart
 	name = "Grape Tart"
@@ -120,7 +120,7 @@
 	        /obj/item/reagent_containers/food/snacks/grown/grapes = 3
 	        )
 	result = /obj/item/reagent_containers/food/snacks/pie/grapetart
-	subcategory = CAT_PIE
+	category = CAT_PIE
 
 /datum/crafting_recipe/food/meatpie
 	name = "Meat Pie"
@@ -131,7 +131,7 @@
 		/obj/item/reagent_containers/food/snacks/meat/steak/plain = 1
 	)
 	result = /obj/item/reagent_containers/food/snacks/pie/meatpie
-	subcategory = CAT_PIE
+	category = CAT_PIE
 
 /datum/crafting_recipe/food/plumppie
 	name = "Plump Pie"
@@ -140,7 +140,7 @@
 		/obj/item/reagent_containers/food/snacks/grown/mushroom/plumphelmet = 1
 	)
 	result = /obj/item/reagent_containers/food/snacks/pie/plump_pie
-	subcategory = CAT_PIE
+	category = CAT_PIE
 
 /datum/crafting_recipe/food/pumpkinpie
 	name = "Pumpkin Pie"
@@ -151,7 +151,7 @@
 		/obj/item/reagent_containers/food/snacks/grown/pumpkin = 1
 	)
 	result = /obj/item/reagent_containers/food/snacks/pie/pumpkinpie
-	subcategory = CAT_PIE
+	category = CAT_PIE
 
 /datum/crafting_recipe/food/tofupie
 	name = "Tofu Pie"
@@ -160,7 +160,7 @@
 		/obj/item/reagent_containers/food/snacks/tofu = 1
 	)
 	result = /obj/item/reagent_containers/food/snacks/pie/tofupie
-	subcategory = CAT_PIE
+	category = CAT_PIE
 
 /datum/crafting_recipe/food/xenopie
 	name = "Xeno Pie"
@@ -169,7 +169,7 @@
 		/obj/item/reagent_containers/food/snacks/meat/cutlet/xeno = 1
 	)
 	result = /obj/item/reagent_containers/food/snacks/pie/xemeatpie
-	subcategory = CAT_PIE
+	category = CAT_PIE
 
 /datum/crafting_recipe/food/frenchsilkpie
 	name = "French Silk Pie"
@@ -179,4 +179,4 @@
 		/obj/item/reagent_containers/food/snacks/chocolatebar = 2
 	)
 	result = /obj/item/reagent_containers/food/snacks/pie/frenchsilkpie
-	subcategory = CAT_PIE
+	category = CAT_PIE
