+++ code/modules/food_and_drinks/recipes/tablecraft/recipes_salad.dm
@@ -11,7 +11,7 @@
 		/obj/item/reagent_containers/food/snacks/grown/apple/gold = 1
 	)
 	result = /obj/item/reagent_containers/food/snacks/salad/aesirsalad
-	subcategory = CAT_SALAD
+	category = CAT_SALAD
 
 /datum/crafting_recipe/food/citrusdelight
 	name = "Citrus Delight"
@@ -23,7 +23,7 @@
 
 	)
 	result = /obj/item/reagent_containers/food/snacks/salad/citrusdelight
-	subcategory = CAT_SALAD
+	category = CAT_SALAD
 
 /datum/crafting_recipe/food/fruitsalad
 	name = "Fruit Salad"
@@ -36,7 +36,7 @@
 
 	)
 	result = /obj/item/reagent_containers/food/snacks/salad/fruit
-	subcategory = CAT_SALAD
+	category = CAT_SALAD
 
 /datum/crafting_recipe/food/herbsalad
 	name = "Herb Salad"
@@ -46,7 +46,7 @@
 		/obj/item/reagent_containers/food/snacks/grown/apple = 1
 	)
 	result = /obj/item/reagent_containers/food/snacks/salad/herbsalad
-	subcategory = CAT_SALAD
+	category = CAT_SALAD
 
 /datum/crafting_recipe/food/junglesalad
 	name = "Jungle Salad"
@@ -59,7 +59,7 @@
 
 	)
 	result = /obj/item/reagent_containers/food/snacks/salad/jungle
-	subcategory = CAT_SALAD
+	category = CAT_SALAD
 
 /datum/crafting_recipe/food/monkeysdelight
 	name = "Monkey's Delight"
@@ -72,7 +72,7 @@
 		/obj/item/reagent_containers/food/snacks/grown/banana = 1
 	)
 	result = /obj/item/reagent_containers/food/snacks/soup/monkeysdelight
-	subcategory = CAT_SALAD
+	category = CAT_SALAD
 
 /datum/crafting_recipe/food/oatmeal
 	name = "Oatmeal"
@@ -82,7 +82,7 @@
 		/obj/item/reagent_containers/food/snacks/grown/oat = 1
 	)
 	result = /obj/item/reagent_containers/food/snacks/salad/oatmeal
-	subcategory = CAT_SALAD
+	category = CAT_SALAD
 
 /datum/crafting_recipe/food/validsalad
 	name = "Valid Salad"
@@ -93,7 +93,7 @@
 		/obj/item/reagent_containers/food/snacks/meatball = 1
 	)
 	result = /obj/item/reagent_containers/food/snacks/salad/validsalad
-	subcategory = CAT_SALAD
+	category = CAT_SALAD
 
 /datum/crafting_recipe/food/edensalad
 	name = "Salad of Eden"
@@ -105,7 +105,7 @@
 		/obj/item/reagent_containers/food/snacks/grown/peace = 1
 	)
 	result = /obj/item/reagent_containers/food/snacks/salad/edensalad
-	subcategory = CAT_SALAD
+	category = CAT_SALAD
 
 /datum/crafting_recipe/food/friedrice
 	name = "Fried Rice"
@@ -114,7 +114,7 @@
 		/datum/reagent/consumable/soysauce = 1
 	)
 	result = /obj/item/reagent_containers/food/snacks/salad/friedrice
-	subcategory = CAT_SALAD
+	category = CAT_SALAD
 
 /datum/crafting_recipe/food/friedrice_veg
 	name = "Veggie Fried Rice"
@@ -128,7 +128,7 @@
 		/obj/item/reagent_containers/food/snacks/grown/carrot = 1
 	)
 	result = /obj/item/reagent_containers/food/snacks/salad/friedrice/veggie
-	subcategory = CAT_SALAD
+	category = CAT_SALAD
 
 /datum/crafting_recipe/food/friedrice_meat
 	name = "Meat Fried Rice"
@@ -138,7 +138,7 @@
 		/obj/item/reagent_containers/food/snacks/meat/cutlet = 1
 	)
 	result = /obj/item/reagent_containers/food/snacks/salad/friedrice/meat
-	subcategory = CAT_SALAD
+	category = CAT_SALAD
 
 /datum/crafting_recipe/food/friedrice_shrimp
 	name = "Shrimp Fried Rice"
@@ -148,7 +148,7 @@
 		/obj/item/reagent_containers/food/snacks/fish/shrimp = 1
 	)
 	result = /obj/item/reagent_containers/food/snacks/salad/friedrice/shrimp
-	subcategory = CAT_SALAD
+	category = CAT_SALAD
 
 /datum/crafting_recipe/food/friedrice_supreme
 	name = "Supreme Fried Rice"
@@ -164,4 +164,4 @@
 		/obj/item/reagent_containers/food/snacks/fish/shrimp = 1
 	)
 	result = /obj/item/reagent_containers/food/snacks/salad/friedrice/supreme
-	subcategory = CAT_SALAD
+	category = CAT_SALAD
