+++ code/modules/mob/living/simple_animal/simple_animal.dm
@@ -123,7 +123,6 @@
 	update_simplemob_varspeed()
 	if(dextrous)
 		AddComponent(/datum/component/personal_crafting)
-		AddComponent(/datum/component/cooking)
 	if(music_component && music_path)
 		AddComponent(music_component, music_path)
