+++ yogstation.dme
@@ -560,7 +560,6 @@
 #include "code\datums\components\wearertargeting.dm"
 #include "code\datums\components\wet_floor.dm"
 #include "code\datums\components\crafting\antag.dm"
-#include "code\datums\components\crafting\cooking.dm"
 #include "code\datums\components\crafting\crafting.dm"
 #include "code\datums\components\crafting\guncrafting.dm"
 #include "code\datums\components\crafting\makeshift.dm"
@@ -1884,7 +1883,7 @@
 #include "code\modules\asset_cache\asset_list.dm"
 #include "code\modules\asset_cache\asset_list_items.dm"
 #include "code\modules\asset_cache\assets\contracts.dm"
-#include "code\modules\asset_cache\assets\cooking.dm"
+#include "code\modules\asset_cache\assets\crafting.dm"
 #include "code\modules\asset_cache\transports\asset_transport.dm"
 #include "code\modules\asset_cache\transports\webroot_transport.dm"
 #include "code\modules\atmospherics\environmental\LINDA_fire.dm"
