+++ yogstation/code/modules/crafting/recipes.dm
@@ -97,7 +97,7 @@
 		/obj/item/reagent_containers/syringe = 1
 	)
 	category = CAT_APPAREL
-	subcategory = CAT_CLOTHING
+	category = CAT_CLOTHING
 
 /datum/crafting_recipe/poppy2
 	name = "poppy pin"
@@ -108,7 +108,7 @@
 		/obj/item/stack/rods = 1
 	)
 	category = CAT_APPAREL
-	subcategory = CAT_CLOTHING
+	category = CAT_CLOTHING
 
 /datum/crafting_recipe/ashtray
 	name = "Ashtray"
